+++ libdariadb/storage/page_manager.h
@@ -47,7 +47,7 @@ namespace dariadb{
 			
 			dariadb::storage::ChunksList get_open_chunks();
 			size_t chunks_in_cur_page()const;
-			size_t in_queue_size()const;
+			size_t in_queue_size()const; //TODO rename to queue_size
 
             dariadb::Time minTime();
             dariadb::Time maxTime();